+++ getJSON.php
@@ -6,21 +6,12 @@
  * @author     Sébastien Lucas <sebastien@slucas.fr>
  *
  */
-use SebLucas\Cops\Input\Config;
 use SebLucas\Cops\Input\Request;
 use SebLucas\Cops\Output\JSONRenderer;
 
 require_once __DIR__ . '/config.php';
 
 $request = new Request();
-// @checkme set page based on path info here
-$path = $request->path();
-// make unexpected pathinfo visible for now...
-if (!empty($path) && empty(Config::get('use_route_urls'))) {
-    http_response_code(500);
-    echo 'Unexpected PATH_INFO in request';
-    return;
-}
 
 header('Content-Type:application/json;charset=utf-8');